earthgpt/requirements.txt
# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
accelerate>=0.27.0
peft>=0.8.0
bitsandbytes>=0.42.0

# Data Processing
datasets>=2.16.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Geospatial
rasterio>=1.3.9
shapely>=2.0.0
pycocotools>=2.0.7

# Training & Optimization
deepspeed>=0.12.0
wandb>=0.16.0
tensorboard>=2.15.0

# Utilities
scikit-learn>=1.3.0
scipy>=1.11.0
einops>=0.7.0
timm>=0.9.12
